profileList DB에서 값 받아와서 띄우는 기능 구현 /성준

+++ gigwork/src/main/java/com/example/demo/model/Post.java
@@ -13,8 +13,8 @@ public class Post {
 	private char post_offer_yn;
 	private Date worktime_s;
 	private Date worktime_e;
-	private int lat;
-	private int lng;
+	private Double lat;
+	private Double lng;
 	private char urgent;
 	private String title;
 	private String content;
